requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy
flask-socketio
boto3
dropbox
gevent
gevent-websocket
psycopg2
cloudinary==1.34.0
pymysql
cryptography>=39.0.0,<42.0.0
gunicorn>=20.1.0
python-dotenv
pymysql

